+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Added full support to Image Streamer Rest API version 300:
 - [#146](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/146) Why is Switch the only resource that directly implements #set_scope_uris?
 - [#166](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/166) I3S - Simplify login to i3s through oneview client
 - Give custom exception classes a data attribute for more error context and default message
+- [#174](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/174) I3S - Integration test for Build Plan failing
 
 # v4.0.0